requirements.txt
Django==3.2.3
flake8==3.8.4
djangorestframework
django-filter
django-cors-headers==3.7.0
psycopg2-binary==2.9.3
django-rest-multiple-models==2.1.3
django-environ==0.4.5
azure-storage-blob==12.11.0